Add a pstack on test failure to diagnose alter_table-randomized-test flakiness
This test has had a lot of client-side timeouts on the verification scan.
It's not clear why the scans are occasionally taking 10+ seconds when the
amount of data is relatively small.

This simply makes the test pstack the tablet server before killing it
on test failure. Tested manually by adding an ASSERT_TRUE(false) in the
Verify() function and checking that the test output had the TS stacks.

Change-Id: Idd56276940d4274f0e7c332c08f45df2d44fbd23
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7234
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>
Reviewed-by: Michael Percy <mpercy@cloudera.com>
Tested-by: jenkins