Increase HybridClock's default slack
This increases the slack to the test value (10 secs). When clocks
are so unsynchronized COMMIT_WAIT transactions will likely timeout
but there shouldn't be any other major problems (average error is
usually pretty low, even if the max error isn't).

This patch also includes cleanup in other places where it's no
longer needed to set the hybrid clock flag to true or to increase
the clock slack.

Change-Id: I426a555bdbc276f4ac3552204aec84e18c155d78
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7249
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>
Tested-by: jenkins