client-test: fix flakiness in TestGetTabletServerBlacklist
This test was occasionally fetching the set of replicas for a tablet
before all replicas had been created and reported. This resulted
in the test failing, as it ran out of replicas earlier than expected.

The fix changes the code to loop until all three replicas are present
before proceeding.

Before the patch, TSAN tests usually failed about 3% of the time.
With the patch, I ran 1000 iterations with no failures.

Change-Id: I2705cf3356c5c4c8f3fc830a960de71f3cde69da
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7258
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins