Fix reading of newly-added NULLable columns without defaults
If a NULLable column is added after data is flushed, and no DEFAULT
is specified for the column, we should materialize NULL values
for it. Prior to this patch, we didn't handle this case properly,
and would result in an error at scan time that the data for the
column was missing.

The patch adds a specific unit test, and also enhances the randomized
test to create nullable columns and insert/update NULL values.

Thanks to Zhang Zhen for reporting this bug.

Change-Id: Ieb99e411e8f786bb001f99f6018a36593b32589e
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7262
Reviewed-by: David Alves <david.alves@cloudera.com>
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins