KUDU-304 (part 3): catalog_manager: send alter table RPCs to leader only
This implements a new TSPicker which selects the leader replica of a
given tablet. The triggering of alter now submits only a single task
per tablet, instead of one per replica.

Additionally, the tablet report handling code needed a slight tweak:
it's important to trigger the alter table RPC _after_ the tablet raft
config has been updated in the catalog, or else it may not be able
to find the most recent leader.

Lastly, to work around KUDU-915, I had to change alter table to hold
the schema lock all the way until after it appends its COMMIT message
to the WAL. This prevents the interleaving of messages in the WAL
identified as problematic in KUDU-915.

A new unit test verifies the basic functionality. I also changed the
randomized alter-table test to use replication. The pstack watcher
previously in this test no longer makes as much sense in the
multi-TS scenario, so I removed it.

I ran the randomized test in TSAN 1000 times, and it failed 6/1000.
Most of the failures seem to be related to KUDU-934 or timeouts.
This flaky rate seems reasonable to commit this and then separately
work on fixing the issues.

Change-Id: Idc1c0396c9d83d77570d3d9ed72bfb7d6bac2060
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7274
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins