Trigger remote bootstrap from the consensus queue
The consensus queue on the leader is the primary vehicle for triggering
remote bootstrap of a remote replica. This patch adds stubbed support
for a StartRemoteBootstrap() method to the ConsensusService RPC
interface. It also adds leader-side detection of the need to remote
bootstrap a remote peer, currently only based on that peer returning a
TABLET_NOT_FOUND response to the normal heartbeating / replication
mechanism.

This patch modifies the PeerMessageQueue to support triggering and
construction of remote bootstrap RPC requests and also adds a simple
unit test for the logic of triggering a remote bootstrap request.

Change-Id: I232253afeeca1f5aff08614f85c784280b8f3352
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7275
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins