Refactor remote bootstrap client
The existing API did not provide the expressiveness needed for remote
bootstrap, since a TabletPeer (which owns the TabletStatusListener for a
tablet) would need to be constructed before a TabletMetadata object was
available from the remote. This refactor solves that problem.

In addition, TabletMetadata is changed to support loading a new
SuperBlock into an existing TabletMetadata object. This allows the
TabletMetadata object to be registered as part of a TabletPeer so that
tablet reports can include the information and status of a remotely
bootstrapping replica. Once the remote bootstrap is done, the superblock
is replaced in-place and normal tablet startup is initiated.

Change-Id: I14697ce3a0302899b9cf5e463d508b9edb7e29cb
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7276
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins