csd: do not build or validate by default
CSD validation is onerous because the required maven plugin isn't yet
available in any public maven repos, and Cloudera's private maven repo isn't
in our pom (for obvious reasons). So let's make validation opt-in.

Likewise, Todd suggested we make the entire CSD build opt-in. JD had asked
for this before (as the CSD cannot be built on OSX), so that's two votes in
favor. Note that this only affects top-level behavior; running 'mvn' from
the CSD subdirectory will still work without the opt-in.

After this patch is merged, we'll need to update the Kudu-CSD Jenkins job.

Change-Id: Ia6b1c19c56ac54a0fd8d35fc2be00625f5b1bc3e
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7277
Tested-by: Adar Dembo <adar@cloudera.com>
Reviewed-by: Todd Lipcon <todd@cloudera.com>