KUDU-891. Fix occasional timeout while creating tables
This should fix a flakiness seen in client-test where table creation
would time out.

The bug turned out to be a race condition introduced in
4a98d4b0ce54bda87ff3f8f4f63940a8d998fb63, which changed TabletPeer to
transition state_from BOOTSTRAPPING to RUNNING only after Consensus was
started.

In most cases, Consensus would mark the tablet peer dirty in Start(), but the
heartbeat wouldn't be generated until some tens of milliseconds later (when the
heartbeat thread got scheduled). So, this would give the TabletPeer time to
switch from BOOTSTRAPPING to RUNNING before the tablet report was generated,
and we'd almost always report the RUNNING state to the master.

It's possible, though, that the TabletPeer::Start() thread would get context
switched after starting consensus but before changing the state to RUNNING. In
this case, we could heartbeat the BOOTSTRAPPING state. In that case, the
master would decide that the tablet was stuck bootstrapping, and thus that
the table was still in the process of being created.

client-test hit this problem more than most because it sets the heartbeat
interval to 10ms vs the longer default, increasing the chances that a heartbeat
would be sent between Consensus starting and the state transitioning.

The fix is simple: TabletPeer::Start() has to mark the tablet dirty after
changing its state. In most cases this won't generate a double heartbeat, since
MarkDirty() is idempotent in the case that the heartbeat hasn't been sent yet.

After moving this call to TabletPeer, there's no longer a need to call it in
LocalConsensus, so I removed it from there (and the Consensus interface).

In trying to debug this, I also noticed that TSTabletManager was taking a
read lock in one spot where it should have taken a write lock, and made that
one-line change here too.

There's no new unit test for this since it's a race that's hard to reproduce,
but client-test is currently 8-9% flaky on our test slaves due to this,
and we should know that it's fixed when that drops back down. Additionally,
I locally inserted a sleep before the consensus state change and reproed the bug,
which then was fixed by this patch.

Change-Id: I773cdd63e81b015e1bbc806db2dd8d1f1698c0c2
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7305
Tested-by: Todd Lipcon <todd@cloudera.com>
Reviewed-by: Michael Percy <mpercy@cloudera.com>