thirdparty: delete the cached git hash immediately upon rebuilding
Imagine a sequence of three builds:
1. Thirdparty's git hash is foo. We've never built thirdparty, so we do that
   first and write 'foo' to .build-hash. The rest of the build completes.
2. Thirdparty's git hash is bar. We compare 'bar' to 'foo' and start
   rebuilding thirdparty. Halfway through, the build is aborted. Now
   thirdparty/installed contains a mix of stuff from 'foo' and 'bar', and
   .build-hash is still 'foo'.
3. Thirdparty's git hash is, once again, foo. We compare 'foo' to 'foo';
   they match, so we don't rebuild thirdparty. The rest of the build fails
   because thirdparty/installed is inconsistent.

This actually happened on a (surprisingly) long-lived build slave, so let's
prevent it from happening again by removing .build-hash before running
anything that might touch thirdparty/installed.

A side effect is that if you build locally, notice that your thirdparty is
rebuilding, and interrupt it, you'll need to pass NO_REBUILD_THIRDPARTY=1
into subsequent builds until .build-hash is regenerated. That's annoying,
but worth it to fix this bug.

Change-Id: I6785f5f99095e88c5d22c63450787b282cc04678
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7306
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>