KUDU-926. don't try to reverse-lookup IP addresses on server start
792e559c27e3b6b8f221166575a5222c5a788526 changed the server startup
to try to reverse-lookup their bound IP addresses before reporting
the IPs to the master.

This turned out to add 5 seconds to all of the server start-ups in
my test environment. Upon further thought, it also doesn't really
seem necessary -- if the user is binding to a specific IP, we can
probably assume that IP is externally accessible, and just publish
the IP address when we register it. KUDU-48 will eventually attack
the problem of more complicated network topologies, etc.

This patch removes the reverse lookup but retains the policy of
replacing wildcard addresses with the FQDN.

Change-Id: Iac69ad2cb4b3fdba539ded14b670863f5bb5041a
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7311
Tested-by: jenkins
Reviewed-by: Michael Percy <mpercy@cloudera.com>