Fix locking on WriteTransactionState fields
This adds lock acquisition around WriteTransactionState::row_ops_ and
schema_at_decode_time_. Without these, we would see TSAN errors when
dumping the list of transactions.

This modifies linked_list-test to hit the /transactionz web page for
a tablet during the workload. It also changes the test to run slightly
longer in "fast test" mode, so that TSAN reliably triggered on this
issue. (TSAN does not run "slow" tests).

Change-Id: Ifdfdcb0289d172a125827a768c78ad998a03f87a
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7312
Tested-by: jenkins
Reviewed-by: Michael Percy <mpercy@cloudera.com>