Remove schema from RowwiseRowBlockPB and remove ColumnBlockPB
The 'schema' field in RowwiseRowBlockPB was never used so this removes
it. This also removes ColumnBlockPB which wasn't used either. Even when
we do need to send columns through the wire we don't likely need a new
pb (ColumnBlockPB had the same fields as pre-sidecar RowwiseRowBlockPB
and only the comments on the data formats differed).

Change-Id: I8451b129c4d9b8608b15a9d176c0d0a6d7d76e24
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7314
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins