Initial draft of whitepaper
This is the initial draft of the white paper for publish with the beta release.
We'll likely continue to tweak it -- I have a couple more YCSB experiments running
over night that should give more accurate perf numbers, and would like to include
latency numbers -- but I think this is OK to publish as a first cut.

Change-Id: I31527fad1dd4a69dd1993812e48edf25d8b93e2e
Reviewed-on: http://gerrit.cloudera.org:8080/935
Reviewed-by: Misty Stanley-Jones <misty@apache.org>
Tested-by: Misty Stanley-Jones <misty@apache.org>