[java client] Add a partial row
The C++ client has for a while used a "partial row" in order to contain information about a row.
This concept was missing in the Java client and adding it will have the immediate effect of being
able to represent row keys in a nice way (like C++'s KuduScanner::AddLowerBound).

No new tests were written as this doesn't change any functionality. TestTimeouts had to be modified
because it started hitting timeouts in a different but equally valid way.

Change-Id: I2d9822194155c0451290fb4727eec4d44e41222a
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7335
Tested-by: jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>