KUDU-782: fix the encoded keys and ranges in Java
As stated in the jira above, after the server-side fix was committed none of the Java unit tests
failed. The reason is that it wasn't using the encoded start/stop row keys at all, only passing
predicates.

This patch uses the newly introduced PartialRow in order to specify the lower and upper bounds
on scanners. The building of the row key was then completely removed from ColumnRangePredicates
and that class was deleted. Once this was all fixed, TestScannerMultiTablet started failing as
expected, so it was fixed too.

TODO adopt the KuduPredicate construct from the C++ client

Change-Id: Ic4a66b7d09b4cfccb06a82b950556d1e29762d59
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7336
Tested-by: jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>