Fix race between responding to an RPC and dumping running transactions
This fixes a race occasionally seen in delete_table-test (in an upcoming
patch) between WriteTransaction::ToString() and responding to that
transaction's RPC.

After sending the RPC response, we delete the RPC request object. We'd
already NULLed the request_ pointer in Commit/Abort, but we didn't clear
the ops_ vector. The ops_ end up referring to memory owned by the
request, so we need to also clear out ops_ at the same time.

Change-Id: Ic0d519605cdc4ced750912725be4fe0f89f3bd3c
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7351
Reviewed-by: David Alves <david.alves@cloudera.com>
Tested-by: Todd Lipcon <todd@cloudera.com>