mem_tracker: add and use AnySoftLimitExceeded()
Todd suggested switching the root memory tracker to use tcmalloc's
consumption metric, so that memory consumption is always accurate. We don't
explicitly track every allocation in Kudu today, so this makes sense, though
I still believe that we'll be better off in the long run were we to do that
(i.e. we'll have better introspection and information available in-process).

We can't switch yet, though, because SoftLimitExceeded() isn't hierarchical,
and that forces some unit tests to mock the consumption of the root tracker
directly. That won't be possible when the root tracker uses a metric, so
let's fix this by introducing AnySoftLimitExceeded(), to be used in most
places where soft limit checking is required. This makes it easy for tests
to inject their own trackers (see maintenance_manager-test for an example).

I think AnySoftLimitExceeded() makes structural sense, but there's an
important semantic difference between it and AnyLimitExceeded(): because
soft limit checks are independent events, actually checking multiple soft
limits means a greatly increased chance of the limit being exceeded,
probably more so than is desired. This doesn't happen in production because
soft limit checks only happen on the tablet, server, and root trackers, and
only the last one has a limit defined. But if that were to change, we'd need
to revisit this behavior.

Change-Id: I8d8c5195b7b37d84e9b222b7d2743f1a39afcbb9
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7373
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>