Prefer to pass TypeInfo instead of raw DataTypes where possible
We're currently inconsistent regarding passing the raw DataType
vs the TypeInfo. This patch addresses this.

This is generally helpful for splitting the client type and the
storage type, but would also be helpful when we migrate from
DataType as an enum to something else with more metadata (e.g.
like we need for DECIMAL).

This also removes the type argument completely from
IndexTreeIterator::Create() since it was not being used.

Change-Id: I8cb45f7bdc8e3d9326dfa58b8815d675a32801fc
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7353
Reviewed-by: David Alves <david.alves@cloudera.com>
Tested-by: David Alves <david.alves@cloudera.com>