In TypeInfo, rename type() to physical_type() and add a new type() that reflects the schema type
Most of the times we refer to TypeInfo::type() we're refering to the
storage type, so the first thing this patch does is a rename of the
old TypeInfo::type() to TypeInfo::physical_type(). However there are
a few cases where we genuinely need the type that is referred to in
the schema, so this patch adds a TypeInfo::type() method to be used
in these cases.

Note that these types are still the exact same, a follow up patch
will flip STRING's physical_type() to be BINARY.

Change-Id: I2debccd8aa40fa303d7229c4fa29e1ad988f62fb
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7360
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>