Make STRING be stored as BINARY
This patch flips the switch and makes STRING use BINARY
as a physical_type(), meaning it's encoded and stored
as BINARY from now on. Block metadata still stores STRING.

This patch doesn't change the plain/prefix/dict encoders names
to reduce noise since that change is mechanical. That is done
in the next patch in the series.

Change-Id: I2abbb9efc77b03b7a2ef07d0e5632e2b39bb19c4
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7361
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins