KUDU-809 (part 1): client: add a KuduSchemaBuilder API
This new API is more readable, and also easier for us to evolve as we add more
features to schemas, since it's based on a builder instead of a constructor
with a large number of parameters.

This also provides better error checking, as it returns a Status when an
invalid schema is specified.

This patch creates the new API, changes client-test to use it instead of the
old API, and adds a bunch of unit tests to verify all of the error conditions.

Note that the new API also solves KUDU-737, since it allows for primary key
columns to be specified even if they are not the first columns in the schema.
In the implementation, we don't support this yet, but the new API will allow
us to later add the feature without breaking existing users.

Change-Id: I1dd8402e7c9bfe7a9c007f6cc0c4f8da5ca9e8d9
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7358
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins