Tombstone tablets when remote bootstrap fails
If remote bootstrap fails due to a failure of the remote, the TS tablet
manager should simply tombstone the tablet locally so it can be retried
later.

We add a test for this, as well as the case where the TS crashes while
remote bootstrapping.

Also fixed a bug in the ExternalMaster that caused Restart() not to
restart on the same port.

Change-Id: I5e23f0e03b90a73922e4ec5b4189ae8f7d93d574
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7399
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins