mem_tracker: switch root tracker to observe tcmalloc memory usage
It is now safe to change the root tracker to report tcmalloc's usage. To
accomplish this, I repurposed the existing FunctionGauge-based tracker logic.
Originally I also wanted to reuse the code in tcmalloc_metrics.cc, but that
didn't work because the root tracker is constructed statically, without any
access to metric entities, and often before the registry is built. So I
copied some of that logic, and as there are no FunctionGauge-based trackers,
right now, I simplified it slightly for this use case.

How does this work? The root tracker will ask tcmalloc for the heap size
with any call to Consume(), Release(), or UpdateConsumption(). I think the
existing Consume()/Release() calls added for tracking various objects should
suffice for reasonable accuracy, but if not, we can add explicit calls to
UpdateConsumption().

It wasn't clear whether to look at tcmalloc's generic.heap_size or
generic.current_allocated_bytes. The latter purports to include tcmalloc
overhead and memory fragmentation, but experimentation revealed that its
value never drops, presumably because as a "heap" size, it makes sense for
it to measure virtual memory.

Change-Id: Ic999543a49f87b1f7d0b4c29a0dcd73d06476cb6
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7374
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>