KUDU-877: offload hole punching from destructor thread
It's not clear why the underlying fallocate() used to punch holes can be so
slow, but let's get it off the LogBlock destructor path.

To do this, there's now a generic way to perform asynchronous work for a
given disk. It's quite simple: each disk has its own singleton thread pool
with an unbounded queue size. Submitted work runs in parallel where
possible, with a fallback to synchronous work if need be. Besides hole
punching, it's also used to parallelize open.

Change-Id: I47ffa98e91119f89d62022ad81b276596a9476aa
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7372
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>