KUDU-861 (part 1): convert C++ alter table API to be more builder-like
This changes the alter table API to look more like the create table API.
In particular, the API now supports "alter column" in a more fluent
syntax, and "add column" reuses the same syntax as creating a new table.

In addition, changed the KuduTableAlterer to return pointers instead of
references per KUDU-944.

This does not yet make the necessary wire protocol and master side changes
to actually support changing encodings or defaults with AlterColumn.
That will come in a later patch.

Change-Id: I19473854c1b7f5c7d612d28ef65fff33e90b3fed
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7447
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins