encoded_key-test: force benchmark to not inline code under test
Before, small changes in the size of the EncodeKey function would cross
the inlining threshold and cause big swings in performance. In particular,
Jingkai provided an optimization which added a few instructions to the
function. The longer function performs significantly better, but because
it was longer, caused gcc to stop inlining it, making the benchmark slow
down. Clang continued to inline it in either case, so the optimization
seemed good on clang and bad on gcc.

This patch forces the benchmark to not inline the function being tested,
so that the speedup is observed consistently.

Change-Id: I521f9a65e6ad261fb366807619bbb534b5bc2737
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7375
Tested-by: jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>