Add an integration test that tests all types
This patch introduces a new test that creates, inserts into and verifies tables
using all possible types. In order to make sure key encoding is tested the
test creates tables with every possible non-composite type (but it's setup in
such a way that we could easily add composite keys in the future).

Change-Id: Ie0d7fdb84088d76310472248f571ed2f55c394fe
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7388
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: Todd Lipcon <todd@cloudera.com>