KUDU-942. Disable core dumps when crashing due to fault injection
This adds a simple call to setrlimit to disable core dumps when crashing due to
an injected fault. This should prevent spewing core files to the disk in tests
where we inject faults and repeatedly restart servers (for example,
RaftConsensusITest.InsertWithCrashNodes).

I manually tested this by running the above-mentioned test with `ulimit -c
unlimited`, and ctrl-Cing the test case just after the fault was injected.
Without this patch, I saw a core file. With it, I didn't.

Change-Id: I5bb340942ff11b91aa682588ff1c90b4da1ab9d9
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7382
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins