KUDU-956 (part 1): suppress ThreadRestrictions error on Client shutdown
This adds a ThreadRestrictions suppression during the client shutdown process.
The client can shut down on the reactor thread in the case that an in-flight
RPC holds the last reference to the client. This then ends up waiting on the
DNS resolution thread pool, which fires the thread restrictions assertion.

Fixing this is likely tricky, and it's causing a lot of test flakiness, so
this commit just adds a suppression for the time being. It shouldn't be critical
since we only hit this case while a client is shutting down, and therefore
we shouldn't care about any latency being introduced.

Change-Id: I26a163c8ba7a2286506e323acab84b9da37bdc50
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7390
Reviewed-by: David Alves <david.alves@cloudera.com>
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins