benchmarks: remove memrowset_throttle_mb flag from benchmark
This flag no longer exists, as throttling is done at the service level.

We should probably modify mt-tablet-test a bit so that it enforces some
throttling of its own -- otherwise the benchmark runs may end up using a lot of
memory. But, this temporary fix will unblock the rest of the benchmarks.

Change-Id: I062885402ec99c10a64b1a86b6eaf7e69967c08a
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7391
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: Todd Lipcon <todd@cloudera.com>