raft_consensus-itest: shorten TestAutoCreateReplica runtime, increase its timeout
This should de-flake the test case, which is currently timing out about 8% of
the time in ASAN builds. Looking at the logs, it appears that the test is doing
what we expect, just a bit slow in ASAN. This patch makes it insert half as
many rows (which is still enough to trigger flushes and log rolls) and also
increases the timeout by a factor of four.

Change-Id: I66319c414ca47ddca608cf7876a7dd5027d69d76
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7400
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Todd Lipcon <todd@cloudera.com>