fault_injection: set coredump_filter to not dump entire address space into abrt
On systems where we have 'abrtd' configured, core dumps are piped into the abrt
core processor regardless of the RLIMIT_CORE setting on the process. This
changes the coredump_filter to make the piped core minimal.

This sped up delete_table-test from 120s to 20s on a system with abrt
configured.

Change-Id: Ie740097034a96015ea138710b705310dda5920e5
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7403
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins