KUDU-967 Fix ORDERED scans not retuning ordered results
This fixes a weird corner case where an ORDERED scan would not
return ordered results. The problem originated in the new scan
API that no longer marks projected columns as key columns. When
the server actually performs the scan based on the projection
it fails to mark the key columns as keys and so the MergeIterator
is built with a keyless Schema. Since Schema::Compare() only
compares the key columns, comparisons would always return 0.
This would transform the MergeIterator in just a slower version
of the UnionIterator.

The fix is to build a schema for the projection by going through
all the columns (not just the missing columns, as before) and making
sure we mark the right columns as keys. This also adds a check
to MergeIterator so that it fails when built with a schema with
0 key columns.

Change-Id: I4016d15a044f028225f0b478a515133f31d976cc
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7416
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins