KUDU-1023. Store last logged OpId in superblock when tombstoning
The term of the last received opid should be used for validation when
considering whether to allow remote bootstrap from a given server.
Since a partitioned replica can increment its term independently, we
should not use the current_term stored in the ConsensusMetadata of a
tombstoned tablet for sanity checks against a rogue leader. We should
use the term of the last-received OpId.

After applying this patch, the existing rogue leader tests still pass,
and a new test was added to ensure that the checks are not overly
strict.

Change-Id: I4eda166dae4d63c12c79777864063661054df792
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7420
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>