KUDU-559: limit the amount of memory used by pending transactions
This patch addresses KUDU-559 by establishing a memory limit on each
tablet's pending transactions. The limit is enforced in the
TransactionTracker, as it's used regardless of the replica's role. A
transaction's request size is used as its memory footprint; this is by no
means exact, but it serves as a useful approximation.

Along the way, TransactionTracker's metric handling was rewritten to use
regular AtomicGauges, reducing lock usage on the critical path.

I also tried to make the ServiceUnavailable handling in TabletService a
little cleaner. This included changing two AddRpcSidecar() calls to CHECK as
they can't really fail (at least not today).

Change-Id: I1b963e29ef72e704a6c3f908b7747895bd60fa6b
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7444
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>