raft_consensus-itest: make the memory pressure test less flaky
On some release builds, 10000 ops isn't enough to consume 8 mb of memory.
We could send more ops, but that increases the test runtime, so let's halve
the memory limit instead. We need to be careful not to go too far, otherwise
the tserver ends up in a nonstop unproductive flush loop.

Let's also disable transaction memory tracking to ensure that the error
message remains consistent.

Change-Id: I09ca2eeab4126488d82d196b0a148c42431d6779
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7465
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>