client: define exported library version
Why a separate version for the library? Because we want to allow it to
evolve independently from the rest of Kudu. For example, we might want to
change Kudu's major version from 1 to 2 because of some huge new feature. If
that version was also the client library's version, it'd appear to be an ABI
break to the library's users, regardless of whether there was a break or not
(and if that new feature was implemented as a set of entirely new classes,
there's no reason for it to break the ABI).

Separately I've looked into symbol versions, which help deal with situations
where a semantic change must be made to a symbol. Without symbol versions,
such a change is breaking and should lead to a major version bump. With
symbol versions, the symbol is instead duplicated, with each copy having its
own version. At runtime, applications link against whichever symbol they
were exposed to at build time. Thus, symbol versions allow breaking changes
to be made without actually breaking backwards compatibility. However,
they're not portable and difficult to maintain, and as we don't actually
need them yet, I've deferred the implementation to later.

I also snuck in a change to make client_samples-test cleanup more robust.

Change-Id: Ie1d9ee6e4fb4d6a095f457103b7f48b77db42193
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7471
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins