transaction_tracker: fix full stack integration tests
These tests hammer the same tablet with dozens of clients, blowing out the
transaction memory limit. On a1228 (spinning disk) and on my laptop (ssd) I
managed to pass the test reliably once I had increased the limit to 64 mb.
As the throttling is only intended to protect against pathological behavior
when there's a live leader with a downed replica majority, it's not terribly
important for the default limit to be low, just for it to exist. So let's
change it to 64 mb.

Change-Id: I6ccecbe7e6e1ad56565c80ba33f945d9d833263b
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7482
Tested-by: Adar Dembo <adar@cloudera.com>
Reviewed-by: Todd Lipcon <todd@cloudera.com>