KUDU-994. Work around MemTracker lifetime issues
The build's flaky now that the MemTracker for the Tablet is tied to the
Tablet object's lifetime and the Tablet may be replaced by a remote
bootstrap. This is a workaround for that issue, which simply adds a
static counter to the name of the created tablet MemTracker each time we
create a new Tablet object.

Change-Id: I130ab728723dc86fd4a1e3f8fcd0632d3811fb81
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7486
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>