KUDU-994: orphan memtracker when primary owner is destroyed
This patch resuscitates the UnregisterFromParent() functionality of the
MemTracker removed in commit 3965101 and uses it to orphan a tracker when
its "primary" owner is destroyed. Here is a problematic sequence of events:
1. Object foo creates a tracker with id "foo".
2. Object bar takes a reference on that tracker. It is OK for bar to
   outlive foo.
3. Object foo is destroyed and recreated.
4. The original tracker lives on through bar's reference, and the attempt
   to create it again when foo is recreated fails.

Even though parent tracker references are only taken in certain places,
there's always the possibility of someone using MemTracker::ListTrackers()
to acquire a reference to every tracker. So, let's fix this. The idea is for
step #3 to orphan the MemTracker so that it can still be used for
consumption operations (which go from the node up), but ignored for
discovery operations (which go from the root down).

To make this work, we need to find all the "primary" tracker owners and have
them explicitly unregister their trackers from the parent when they
themselves are destroyed. It isn't necessary to do this when the owner uses
FindOrCreateTracker(), as reusing an existing tracker is fine too.

I ran DeleteTableTest.TestAutoTombstoneAfterRemoteBootstrapRemoteFails in a
loop on my laptop. Previously it'd fail within 1-2 attempts. Now it hasn't
failed in hundreds of tries.

Change-Id: Iac2bfbb9ab4f5c1291076c2cb881212474b930ef
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7491
Reviewed-by: Michael Percy <mpercy@cloudera.com>
Tested-by: jenkins