KUDU-987 (part 1): Make MvccManager track whether a timestamp is APPLYING yet
This changes the MvccManager from tracking a set of in-flight timestamps to
instead track a map from timestamp to state. A timestamp is either RESERVED
(i.e has been started, but may still be allowed to abort) vs APPLYING
(at which point we're guaranteed that it will proceed to commit in a bounded
amount of time).

The MvccManager also now asserts that we never Abort() an APPLYING transaction,
and we never COMMIT a RESERVED transaction.

This patch changes ScopedTransaction to no longer automatically Commit() in its
destructor. Instead, it Aborts(). This should help us ensure that, once we
call StartApplying(), it's the 'point of no return' beyond which we must
definitely Commit().

Change-Id: Ic0576f9cd7b51168e07b0175ab7d2e984f6b9c78
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7505
Tested-by: jenkins
Reviewed-by: Michael Percy <mpercy@cloudera.com>
Reviewed-by: David Alves <david.alves@cloudera.com>