KUDU-1009. Fix crasher in remote bootstrap service when tablet deleted
If the tablet is deleted while the server is serving a remote bootstrap
to a remote client, it crashes when the session is shut down. This fixes
that issue, cleans up a couple other things, and adds a regression test
that failed before the fix.

This also cleans a somewhat sketchy API issue in TabletPeer. We should
not return a const ref to something that is accessed under a lock and
may be deleted later. Conversely, we should also return a const ref to a
const member instead of a copy when we can safely do that.

Change-Id: Ibfc28e497444ce8093804628aa24d6f0b11697cb
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7506
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>